use HAL_PIXEL_FORMAT_YCrCb_420_SP for HAL_PIXEL_FORMAT_YCbCr_420_888
PD#146294

use HAL_PIXEL_FORMAT_YCrCb_420_SP for HAL_PIXEL_FORMAT_YCbCr_420_888
(Android flexible YCbCr 4:2:0 formats)

This format allows platforms to use an efficient YCbCr/YCrCb 4:2:0
buffer layout, while still describing the general format in a
layout-independent manner.  While called YCbCr, it can be
used to describe formats with either chromatic ordering, as well as
whole planar or semiplanar layouts.

struct android_ycbcr (below) is the the struct used to describe it.

This format must be accepted by the gralloc module when
USAGE_SW_WRITE_* or USAGE_SW_READ_* are set.

This format is locked for use by gralloc's (*lock_ycbcr) method, and
locking with the (*lock) method will return an error.

When used with ANativeWindow, the dataSpace field describes the color
space of the buffer.

Change-Id: I1167c091350f17f8f6e4477841d0ab485e57954d
Signed-off-by: Jiyu Yang <jiyu.yang@amlogic.com>